reorganizes code
